+++ src/caselawclient/models/documents/versions.py
@@ -26,6 +26,9 @@ class VersionType(Enum):
     EDIT = "edit"
     """ This version has been created as the result of a manual edit. """
 
+    RESTORE = "restore"
+    """ This version has been created as the result of a restore of a previous document. """
+
 
 class VersionAnnotation:
     """A class holding structured data about the reason for a version."""